Use the default package manager and pip executable of the OS
This change is needed of CentOS8. The RDO CentOS8 image
comes with 'yum' package not installed, as 'dnf' is the
default package manager.
We take advantage of the ansible fact 'ansible_pkg_mgr'.

Also, we have to specify explicitly the version of pip
binary.

Also adjust repoquery command for CentOS8 as it doesn't
honor --plugins option.

Change-Id: I00c861f5a9c298620955ca85fd70a41899702a0d
Depends-ON: https://review.rdoproject.org/r/#/c/24104/